Updated uncompressedSize and compressedSize properties of UZKFileInfo to match the 64-bit version of the file info (Issue #37)

+++ Source/UZKArchive.m
@@ -1739,7 +1739,7 @@ compressionMethod:(UZKCompressionMethod)method
 }
 
 
-- (NSData *)readFile:(NSString *)filePath length:(NSUInteger)length error:(NSError * __autoreleasing*)error {
+- (NSData *)readFile:(NSString *)filePath length:(unsigned long long int)length error:(NSError * __autoreleasing*)error {
     if (![self openFile:error]) {
         return nil;
     }
